+++ src/main.ts
@@ -27,6 +27,7 @@ import {
   isScalar,
   isStructType,
   isStructTypeName,
+  isTimeOfDay,
   isTimestamp,
   isValueType,
   isWholeNumber,
@@ -37,6 +38,7 @@ import {
   toReaderCall,
   toTypeName,
   valueTypeName,
+  wrapperTypeName,
 } from './types';
 import SourceInfo, { Fields } from './sourceInfo';
 import {
@@ -75,6 +77,8 @@ import { generateSchema } from './schema';
 import { ConditionalOutput } from 'ts-poet/build/ConditionalOutput';
 import { generateGrpcJsService } from './generate-grpc-js';
 import { generateGenericServiceDefinition } from './generate-generic-service-definition';
+import { generateNiceGrpcService } from './generate-nice-grpc';
+import { join } from 'path';
 
 export function generateFile(ctx: Context, fileDesc: FileDescriptorProto): [string, Code] {
   const { options, utils } = ctx;
@@ -150,6 +154,8 @@ export function generateFile(ctx: Context, fileDesc: FileDescriptorProto): [stri
       fileDesc,
       sourceInfo,
       (fullName, message, sInfo, fullProtoTypeName) => {
+        // console.dir({ fileDesc, sourceInfo }, { depth: 8 });
+        // console.dir({ fullName, message, sInfo, fullProtoTypeName }, { depth: 7 });
         const fullTypeName = maybePrefixPackage(fileDesc, fullProtoTypeName);
 
         chunks.push(generateBaseInstanceFactory(ctx, fullName, message, fullTypeName));
@@ -171,6 +177,9 @@ export function generateFile(ctx: Context, fileDesc: FileDescriptorProto): [stri
         if (options.outputPartialMethods) {
           staticMembers.push(generateFromPartial(ctx, fullName, message));
         }
+        staticMembers.push(generateGetTimestampKeys(ctx, fullName, message));
+        staticMembers.push(generateGetTimestampKeysValue(ctx, fullName, message));
+        staticMembers.push(generateGetMessageKeys(ctx, fullName, message, fileDesc));
 
         staticMembers.push(...generateWrap(ctx, fullTypeName));
         staticMembers.push(...generateUnwrap(ctx, fullTypeName));
@@ -219,6 +228,8 @@ export function generateFile(ctx: Context, fileDesc: FileDescriptorProto): [stri
       uniqueServices.forEach((outputService) => {
         if (outputService === ServiceOption.GRPC) {
           chunks.push(generateGrpcJsService(ctx, fileDesc, sInfo, serviceDesc));
+        } else if (outputService === ServiceOption.NICE_GRPC) {
+          chunks.push(generateNiceGrpcService(ctx, fileDesc, sInfo, serviceDesc));
         } else if (outputService === ServiceOption.GENERIC) {
           chunks.push(generateGenericServiceDefinition(ctx, fileDesc, sInfo, serviceDesc));
         } else if (outputService === ServiceOption.DEFAULT) {
@@ -301,7 +312,8 @@ export type Utils = ReturnType<typeof makeDeepPartial> &
   ReturnType<typeof makeTimestampMethods> &
   ReturnType<typeof makeByteUtils> &
   ReturnType<typeof makeLongUtils> &
-  ReturnType<typeof makeComparisonUtils>;
+  ReturnType<typeof makeComparisonUtils> &
+  ReturnType<typeof makeNiceGrpcServerStreamingMethodResult>;
 
 /** These are runtime utility methods used by the generated code. */
 export function makeUtils(options: Options): Utils {
@@ -314,6 +326,7 @@ export function makeUtils(options: Options): Utils {
     ...makeTimestampMethods(options, longs),
     ...longs,
     ...makeComparisonUtils(),
+    ...makeNiceGrpcServerStreamingMethodResult(),
   };
 }
 
@@ -421,9 +434,9 @@ function makeByteUtils() {
       const btoa : (bin: string) => string = ${globalThis}.btoa || ((bin) => ${globalThis}.Buffer.from(bin, 'binary').toString('base64'));
       function base64FromBytes(arr: Uint8Array): string {
         const bin: string[] = [];
-        for (const byte of arr) {
+        arr.forEach((byte) => {
           bin.push(String.fromCharCode(byte));
-        }
+        });
         return btoa(bin.join(''));
       }
     `
@@ -629,6 +642,19 @@ function makeComparisonUtils() {
   );
 
   return { isObject, isSet };
+}
+
+function makeNiceGrpcServerStreamingMethodResult() {
+  const NiceGrpcServerStreamingMethodResult = conditionalOutput(
+    'ServerStreamingMethodResult',
+    code`
+      export type ServerStreamingMethodResult<Response> = {
+        [Symbol.asyncIterator](): AsyncIterator<Response, void>;
+      };
+    `
+  );
+
+  return { NiceGrpcServerStreamingMethodResult };
 }
 
 // Create the interface with properties
@@ -1437,6 +1463,76 @@ function generateToJson(
     }
   });
   chunks.push(code`return obj;`);
+  chunks.push(code`}`);
+  return joinCode(chunks, { on: '\n' });
+}
+
+function generateGetTimestampKeys(ctx: Context, fullName: string, messageDesc: DescriptorProto): Code {
+  const { options, utils, typeMap } = ctx;
+  const chunks: Code[] = [];
+
+  const timestampKeys: string[] = [];
+
+  chunks.push(code`
+    __SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED_getTimestampKeys(): string[] {
+  `);
+  // add a check for each incoming field
+  messageDesc.field.forEach((field) => {
+    const fieldName = maybeSnakeToCamel(field.name, options);
+    if (isTimestamp(field) || isTimeOfDay(field)) {
+      timestampKeys.push(`'${fieldName}'`);
+    }
+  });
+  chunks.push(code`return [${timestampKeys.join(',')}];`);
+  chunks.push(code`}`);
+  return joinCode(chunks, { on: '\n' });
+}
+
+function generateGetTimestampKeysValue(ctx: Context, fullName: string, messageDesc: DescriptorProto): Code {
+  const { options, utils, typeMap } = ctx;
+  const chunks: Code[] = [];
+  chunks.push(code`
+      timestampKeys:[] as string[]
+  `);
+  return joinCode(chunks, { on: '\n' });
+}
+
+function generateGetMessageKeys(
+  ctx: Context,
+  fullName: string,
+  messageDesc: DescriptorProto,
+  fileDesc: FileDescriptorProto
+): Code {
+  const { options, utils, typeMap } = ctx;
+  const chunks: Code[] = [];
+
+  const messageKeys: string[] = [];
+  chunks.push(code`
+  __SECRET_INTERNALS_DO_NOT_USE_OR_YOU_WILL_BE_FIRED_getMessageKeys(): {[key:string]:string[]} {
+  `);
+
+  // add a check for each incoming field
+  messageDesc.field.forEach((field) => {
+    // messageKeys.push(`'${join(fileDesc.package, fileDesc.name)}'`);
+    const fieldName = maybeSnakeToCamel(field.name, options);
+    const isWrapper = wrapperTypeName(field.typeName);
+    if (
+      !isTimestamp(field) &&
+      !isTimeOfDay(field) &&
+      !isWrapper &&
+      !isPrimitive(field) &&
+      !isEnum(field) &&
+      !field.typeName.includes('google')
+    ) {
+      messageKeys.push(field.typeName);
+    }
+  });
+
+  chunks.push(
+    code`return ${JSON.stringify({
+      [['', fileDesc.package, messageDesc.name].join('.')]: messageKeys,
+    })}`
+  );
   chunks.push(code`}`);
   return joinCode(chunks, { on: '\n' });
 }